Update futurium_addon.py
Added square topology button, for brick and carpet meshes. Currently not suitable for top slice. Currently has some issues, will need to find a better way of selecting faces that need cutting but mostly works for now

Updating house exporter. Currently exposes a list of naming of which empties can be targeted so child objects of it, are not included in resetting/freezing transforms. At the moment, when brought into Maya, these nodes/empties take on the 100 scale and child objects have .001 applied. Works out but this bug will be fixed when I figure it out.

Also added filepath the user can input themselves